refactor: improve UI styling and fix color types
Simplify the recording control UI layout in App.tsx and fix implicit integer conversion issues in TimelineTheme.kt by adding explicit .toInt() calls.

+++ android/app/src/main/java/com/litecut/app/timeline/TimelineTheme.kt
@@ -15,12 +15,12 @@ object TimelineTheme {
     // Core Colors matching React App (#0c0c0e)
     var backgroundColor: Int = 0xFF0C0C0E.toInt()         // Dark canvas background
     var trackBackgroundColor: Int = 0xFF0C0C0E.toInt()    // Subtle track lane background
-    var trackSeparatorColor: Int = 0x0AFFFFFF             // 3% opacity white border line between lanes
+    var trackSeparatorColor: Int = 0x0AFFFFFF.toInt()     // 3% opacity white border line between lanes
     var headerBackgroundColor: Int = 0xF20C0C0E.toInt()   // Time ruler background (95% opacity)
-    var headerBorderColor: Int = 0x0DFFFFFF               // Border line below ruler
+    var headerBorderColor: Int = 0x0DFFFFFF.toInt()       // Border line below ruler
 
     // Ruler Grid Colors
-    var rulerTickColor: Int = 0x33FFFFFF                  // Tick marks color
+    var rulerTickColor: Int = 0x33FFFFFF.toInt()          // Tick marks color
     var rulerTextColor: Int = 0x88A1A1AA.toInt()          // Label colors (zinc-400)
     var rulerTextSize: Float = 10f                        // Text size for ruler times (10sp)
 
@@ -45,10 +45,10 @@ object TimelineTheme {
     var clipImageSelEndColor: Int = 0xFF065F46.toInt()
 
     // Borders
-    var clipBorderColorVideo: Int = 0x336366F1
-    var clipBorderColorAudio: Int = 0x33A855F7
-    var clipBorderColorText: Int = 0x33F59E0B
-    var clipBorderColorImage: Int = 0x3310B981
+    var clipBorderColorVideo: Int = 0x336366F1.toInt()
+    var clipBorderColorAudio: Int = 0x33A855F7.toInt()
+    var clipBorderColorText: Int = 0x33F59E0B.toInt()
+    var clipBorderColorImage: Int = 0x3310B981.toInt()
 
     var selBorderColorVideo: Int = 0xFF60A5FA.toInt()
     var selBorderColorAudio: Int = 0xFFC084FC.toInt()
